32_code/32_Software/requirements.txt
"""
robotNav依赖包列表
"""

# 串口通信
pyserial>=3.5

# 数值计算和信号处理
numpy>=1.21.0
scipy>=1.7.0

# SLAM定位建图
breezyslam>=0.5.0

# 可视化（可选）
matplotlib>=3.3.0
Pillow>=8.0.0

# 测试（可选）
pytest>=6.0.0
pytest-cov>=2.10.0
